Reestructuración del modulo inventario e Independizació de la nueva clase Stock

+++ Data/Modelos/ProductoPorPeso.cs
@@ -9,7 +9,7 @@ namespace IlkaPoint.Data.Modelos
     // esta clase hija es cuando en el inventario hay productos que se venden por peso en kilogramo, libras, etc.   
     public class ProductoPorPeso : Producto
         {
-            public float peso { get; set; }
+            public decimal peso { get; set; }
             public override decimal calcularPrecio()
             {
             return precio * (decimal)peso;
